Fixed MSG instruction
It is no longer needed to supply 'this' argument via arguments list and
in the MSG instruction line - 'this' is only included in the args list.
This will prevent certain class of errors that could emerger when
different object was placed in args and different object supplied to MSG
instruction.