Change type used in MSG instruction to Type*
This change was needed because not all Viua-native types that have
foreign methods registered are derived from Object class, e.g. String is
not but it's foreign methods are available to user code.

This also does away with static_cast or dynamic_cast because registers
and arguments hold Type pointers.