Remove hlas.asm file with asm-syntax suggestions
Nearly all ideas described in this file have been implemented.
Some with slightly modified syntax, some with more expressive power.

One thing that hasn't been implemnted is vector generation, but it can
be emulated with available tools.
The code below:

    vpush (vpush (vec 1) (istore 2 0)) (istore 2 1)
    print 1

would print:

    [0, 1]

While the syntax is ugly, it gets the job done.
But it's probably not the way you'd like to generate vectors anyway.