CPU checks if main thread has stopped
This should fix a bug when an exception has been thrown in main thread
but the CPU did not check for unhandled exceptions and
just kept executing the thread leading to segmentation faults as this
resulted in certain exceptions being rerun with nulled pointers etc.

Unhandled exceptions are no fun.