Threads store pointers to their parents
Now, `main/1` will not complain about about orphaning threads that are
not its direct children (e.g. spawned by a non-main detached thread).

After a new thread is launched, Viua attaches it to a parent, which is
the thread that launched it.
Then, the thread continues execution and can be either joined or
detached.

If a thread is detached it becomes "parentless" and receives status
similar to main thread; since Viua threads are independent of each other
and contain all their environment a detached thread may, for example,
continue execution even after `main/1` function exits.

If a thread is not detached and it's parent exits, VM will abort
execution.
This is not yet fully enforced (as of this commit) but will be
implemented; now, only if main exits and it has some attached threads
the VM will abort.