Make watchdog_process nullptr to prevent VM confusion after termination
If the watchdog process is terminated by a runaway exception the VM
attempts to resurrect it.
Since there can be only one watchdog the VM checks if the watchdog_process
pointer and if it is not set to nullptr it throws an exception to signal
that another watchdog process is already running.

The watchdog_process pointer must be set to nullptr before watchdog is
resurrected to prevent the machine from complaining about watchdog being
"already spawned" when it just died.