Fix memory leak from terminated watchdogs
Processes do not delete the object that has been thrown because it will
be either transferred to parent process, or reported as a part of stack
trace summary after the program crashes, or sent to a watchdog process
if there is one.
Since watchdog death messages are neither sent to watchdog, nor passed
to any other thread the VM should just delete the exception that cause
the watchdog to terminate.