Fix: assembler refuses to link duplicate symbols
Duplicate symbols are detected during static linking.
The assembler complains with following error message:

> duplicate symbol 'S' found when linking 'M'

Which means that the assembler found symbol 'S' in module 'M', but
that symbol was already defined be a previously linked module.

When such a situation arises, you should inspect the command line that
produced the error and check if any module that comes on it before the
module 'M' contains symbol 'S'.