Allow functions to specify variable arity
Assembler can be informed that a function has variable arity by ending
its name with just '/' character.
For example:

- function 'foo' has undefined arity (which is an error),
- function 'foo/1' has defined arity and it is 1,
- function 'foo/' has variable arity,

Functions with variable arity generate a warning when --Wundefined-arity
option is passed to assembler, but do not generate error even if
--Eundefined-arity option is present.
In case of functions with variable arity this warning may be treated
more like a diagnostic message telling you "your code may be improved".