Suspend a process before requesting FFI call
This fixes a race condition that would hang the VM when the following
series of events occured:

- code requests FFI call,
- a request is sent to FFI scheduler,
- OS task scheduler starts executing Viua FFI scheduler,
- FFI call finishes and wakes requesting process up,
- normal Viua scheduler suspends the requesting process,

Normal and FFI scheduler run on different thread, thus the FFI call
could actually return before the requesting process was suspended and
the "wakeup" message was lost and the process remained suspended forever.

This is now fixed as the process is suspended before the FFI call
request is sent.