Processes receive pointer to their scheduler
This is required if the processes are to spawn new processes as
they have to *somehow* tell the scheduler to create a new process.

This will also enable the next improvement: CPU access will be
routed through the scheduler, which will allow decoupling process
code from the CPU code and further parallelisation of the VM.
Fewer locks will be required as only schedulers will "talk" with the
heart of the VM, processes will only access information local to their
scheduler (and since they run sequentially under schedulers control
the access can be lock-less).