Mark watchdog as "hidden" process
For a lack of better name.
In this context "hidden" means a process that can be only directly sent
a message, i.e. it has no mailbox in the CPU.

Since watchdogs are scheduler-specific it makes sense to avoid the trip
to the CPU (and associated locking) to transfer a message to a watchdog
process.
However, watchdogs are user-supplied so they fetch messages using normal
receive instruction.
Thus, watchdogs must be specially marked - and this marking is named
"hidden" for now.