Remove return type specifiers from arithmetic and logic instructions
These "return type specifiers" told the VM what type the result should
be, and to what type the operands should be converted before either
arithmetic or logic operation took place.
This proved unwieldy - it made ALU instructions a special case in
syntax.
Also with the advent of simplified types for arithmetic type
specifiers lost their weight because there are not only two types:
integer and float.

The semantics changed to:

- an ALU instruction takes two input values: lhs and rhs
- rhs value is converted to the same type as lhs value
- operation is executed
- if it was arithmetic operation, the result type is the same as the
  type of the lhs value
- if it was logic operation, the reslt type is boolean