Fix prevent data-race between destructor and fetching return value
Marking a process as "joined" signals to the scheduler that the process
and its data can be discarded.
A race condition was present in code when the process was first marked
as joined, and after that return value was fetched.
This could sometimes result in a read from deleted object.