Merge thread-hardened code into main branch
The thread-hardening will continue on a separate branch.
It seems I have to look deeper and switch to manual to find the
remaining races as TheadSanitizer is can find them, but is unable to
restore stack traces for me to analyse.
Yay for by-hand bug-hunting I guess...

Also, some TSan does not *always* detect a race condition.
The number of warnings reported varies (by one or two entries but
still).
So I will also have to run the test suite a few times after the second
round of fixing is finished to see if there's anything left to do.