Add basic SA for blocks
This is able to analyse "enter" and "leave" instructions. Entered blocks
share their register usage profile with their surrounding frame (and
this is implemented).

Error messages are better, but may still be misleading (e.g. they can
report an unused value in function "foo/1", when the value was in fact
defined inside a block "bar"). This is helped a bit by the fact that the
error trace includes line and character number, so it will be obvious
inside which block the error originates (or rather - it *should* be
obvious).