Fix error rethrowing on false branch
The error should only be rethrown if the error is different than the one
thrown by the true branch, which in practice means that the error must
be an "unused value" error but for a different register.

An example: if the true branch leaves register 1 unused and the false
branch leaves register 2 unused - that's OK, but if they both leave
register 1 unused - that's not and the error must be rethrown.