Add two scripts for background compilation
I usually open a tmux session, run './scripts/watch.sh' in it and
detach.
Whenever I write a file it triggers a compilation and pushes a desktop
notification if the change compiles or not.
For me it's useful, so I include the scripts in the repository.
Example shell session:

    (not tmux) ]$ tmux new -s viuacc
    (in  tmux) ]$ ./scripts/watch.sh
                  *C-b d*

And I happily hack away.