A test for one more corner case of saturating arithmetic
This time the corner case was: (X - MINIMUM).
Example: (-1) - (-128).

The result should be 127 as it is perfectly in range for 8-bit integers
but due to all operations being saturating (and a wrap-around effect)
taking the two's complement of -128 got us 127 instead of 128.

And since `(-1) - (-128)` is not the same as `(-1) - (-127)` the result
returned in this case was incorrect.
This was fixed by incrementing the final result if the right hand side
of the saturating signed subtraction is the minimum value.