Allow move instruction to set parameters
The following code is now valid:

    text %1 local "Hello World!"  ; create a text value in register 1
    frame %1                      ; create a frame with 1 argument slot
    move %0 arguments %1 local    ; move the text as the argument

This is the first step towards deprecating the `pamv`, `param`, and
`arg` instructions in user code. Their functionality will be shifted to
`move` and `copy` instructions - teh code will become more readable and
clear as a result.
The deprecated instructions will be retained as an implementation
detail: it will be faster to switch on the opcode then add more cases to
paths of `copy` and `move` instructions.