Allow adding comments to errors
Comments (not to be confused with aside notes) will look like this:

    blah.asm:42:42: error: some error
    blah.asm:42:42: note: some note

         40 |    move %1 local %0 parameters
         41 |    move %2 local %1 parameters
    >>>> 42 |    move %3 local %1 parameetrs
            |                  ^~ ~~~~~~~~~~
            |                  ^ some aside note
            |
            > a comment
            > some more comments
            |
         43 |    ptr %4 local %2 local
         44 |    atom %5 local 'foo'

Comments will be provided to add a bit of, well, commentary to an error.
Be it an excerpt from the manual, an explanation of what went wrong, a
URL to a page describing the cause of the error in detail, code
examples, etc.
Comments are designed to include longer pieces of text that will fit in
neither notes (which should be used for shorter explanations), nor aside
notes (which should be used for *really* short text, e.g. "did you mean
...?" like messages).