Use value wrappers instead of unique "type of value" for every register
This will help with implementing type inference, as the type of one
value will be a part of another value - by virtue of being a pointer to
an inside of the structure representing that type. If all types are
represented as such wrappers then implementing inference is
straightforward.

Assume we have the following in a local register 1:

    vector#0 of value#1

and use the `vat` instruction, to extract an element to local register
2; then the value in local register 2 will have the following type:

    pointer#2 of value#1

This `value#1` is the same structure as in the first line due to the
fact that types are sharing parts. Then, a few instructions further, we
use the `iinc` instruction on the value in local register 2. Thus, we
can deduce its type to be `pointer of integer`, and update the type
accordingly:

    pointer#2 of integer#1

As you can see, the "id" of the type did not change, but its name did.
Since the vector type also uses the same "id", local register 1 is now
known to contain:

    vector#0 of integer#1

Which is what we wanted all along.