Allow rebinding value wrapper to different values
The fact that a value wrapper is just that - a wrapper, and rebinding it
reassigns the value it wraps makes implementing type inference easy. If
we want to deduce the type of elements of a vector let's just copy the
wrapper for the contained type, use some more direct instruction to
deduce it, and just rebind the wrapper used for the "direct"
instruction (e.g. `iinc` which takes integers).
The blank in type of the vector will be "automagically" filled.