Move stringificatio functions higher
They are static because I don't want to declare them (and I will delete
them when they are no longer useful, but the compiler will not be able
to detect that without the `static` keyword). This means that to use
them in the whole file they need to be nearly at the top.