Allow rebinding Value_wrappers to new maps
The index stays the same, but now points to a different map. It is
useful when cloning function states, and when moving them - it is easier
to just clone assigned values and rebind them than to traverse the
hierarchy and rebind each value that should be rebound (although that
would be the correct behaviour).