Analyse if arms independently
This required changing function signatures and plenty of new code, but
hey, now we can actually check if arms! Which is a cool feature but we
had this already - in the old, ugly analyser...

The good news is that we're approaching feature parity with the old
code and already have some more advanced checking in place. All in all
things look pretty much OK.