Import directive works on modules, not files
It is a much better design to specify module's name and let the loader
look for the file than to specify the file's name. As a bonus it forces
a specification of the way modules map to file structre.

HOW TO LINK A MODULE?

    .import: [[static]] Library::Module

This willl look for "Library/Module.module" file in every directory
specified in VIUA_LIBRARY_PATH environment variable.