Add basic support for dynamic imports
Modules specified for dynamic imports:

    .import: [[dynamic]] Module_name

will be loaded by the assembler, and functions exported by them will be
made visible inside the compilation unit. However, no function from a
dynamic module will be emitted in the module or executable produced by
the current compilation unit - as a dynamically linked moduel should.

What is left to be done is implementation of the linking logic at
startup, after a module or executable is loaded by the VM.

For now, you should use the static import feature:

    .import: [[static]] Module_name

for semi-automatic linking.