Fix: stop analysing a control flow path after encountering a tail call
Tail call causes the control flow to be transferred to a far away land
that we have no knowledge of (similarly to what a RETURN or THROW
instructions do). It is not only pointless but also incorrect to
continue analysing the current control flow path as it is broken at the
point of the tail call (as the currently executing frame is replaced).