Fix: allow static linking of more than one module in a single run
This bug was triggered by the compiler project I do as part of getting
my engineer's degree. It's incredibly weird that I have never before
tried to link more than one module in a single invocation of the
assembler... Well, it's not the first time I encounter a weird bug
during the development of the VM. Lucky me that fixing this one took
roughly 10 minutes without first hunting for it for several days.