Fix: only the first instruction inside a function should not be analysed
The first instruction of a function is always ALLOCATE_REGISTERS and is
analysed before the arm-analysing function is entered. The logic
skipping the first instruction was interfering with block analysis. The
flag is now controlling whether the first instruction should be analysed
or not (and is set to false only when the function is first entered).