Allow passing additional arguments to watchdog functions
Watchdog sometimes needs to receive additional information to be able to
fully handle process crash, not only the exception information. This may
be a PID of the supervisor process, a logging identifier, etc.

This commit implements this:

    frame %2
    move %1 arguments %some_value local
    watchdog watchdog_fn/2

Argument slot 0 should always be left empty, as error information is
passed to the watchdog inside it.