Log how many processes were stolen for the scheduler
When a scheduler notices that it has no processes it asks the kernel to
steal some processes on its behalf from another scheduler. Sometimes
there are no processes to steal and the kernel will return none. But in
other cases, let's log how many stolen processes the scheduler received
from the kernel.