Cancel I/O requests automatically when they go out of scope
An I/O interaction *MUST* have a handle existing if is to be carried out
by the I/O scheduler. An I/O interaction may be marked as "fire and
forget" if the handle is to be dropped, but the default is to keep
trying as long as there is anybody interested in seeing the result of
the request and abort otherwise.