Basic write interaction works
With this change in place the most essential I/O (reading from standard
input and writing to standard output) works without foreign modules,
using just VM opcodes.

What is left to be done is creating more interactions then refactoring
common code out (while making provisions for custom interactions that
are not implemented with POSIX I/O system calls). The scheduler should
be rewritten to just fetch (file descriptors, opcode, data) tuples and
wait for several I/O requests in parallel - we are already using the
select(3) call, but wait for just one file descriptor which is terribly
inefficient.

Also, prevent sending I/O request handles between processes for now
since destroying them causes segmentation fault. A better scheme for
aborting and cancelling requests should be devised - an I/O request
being cancelled after it goes out of scope and was not waited for MUST
NOT crash the VM.

Another thing is to prevent several I/O schedulers from fetching I/O
requests for the same file descriptor. We would not want to run several
writes in parallel to avoid corrupting the data. Something like this is
probably needed:

    auto q = std::queue<I/O request*>{};
    auto q_map = std::map<fd, std::queue<I/O request>>{};
    auto fd_grabbed_by = std::map<fd, I/O scheduler*>{};

The 'q' field is used to notify the schedulers about work. A scheduler
then checks the fd associated with the request and checks if it is free
to run it - the condition is true if no scheduler has grabbed the fd, or
if the scheduler that did grab the fd is the scheduler checking. If the
scheduler may run the request it fetches the real request (not just the
pointer); otherwise it checks the next request on the queue.

Maybe reads and writes can be run in parallel? Should it be configurable
when opening the port? The I/O subsystem will definitely grow more
complex...