I/O requests are attached to the kernel, not a process
This avoids the segmentation faults that could have happened if an I/O
request outlived the process that spawned it. When it went out of scope
it tried to automatically cancel or abort itself by going through the
process interface and, since the process was already destroyed, crashed
the whole VM.

Now this will not happen since the kernel outlives any other object.

Sending I/O requests may be useful and attaching them to the kernel
instead of to a process makes this possible.