IO_fd may own the file descriptor or may have just borrowed it
It "owns" the file descriptor if it was used to create it. Owning IO_fd
is responsible for proper shutdown of the file descriptor when the port
goes out of scope.

It "borrows" the file descriptor when it is used as a wrapper for a raw
integer representing one. Borrowing IO_fd is not responsible for the
shutdown of the file descriptor borrowed and is only used as a wrapper
providing I/O interactions.