Fix an ancient bug
Arguments passed to functions were always copied instead of being moved
due to inverted condition the VM was using to determine when to move and
when to copy.

I am *astonished* that it took so long to detect. No less amazing is
that the first non-copyable type (IO_fd) was introduced only recently,
several years after the project started.