Document what each instruction decodes
I need to create an abstract model of what is needed, and then fix the
encoding-decoding of instuctions once and for all (at least until I
again learn a better way to do it). This is related to the "pluggable
decoders" issue 21a6d21083c36978815291ea1e706649d659ac7e.