Another bit of documentation
Yeah, I was pretty lazy with the VM for quite some time... Not that I am
happy about it, but I am not going to cry about this fact, too. It is a
hobby project after all, and I am allowed to take breaks from it.
Nobody's paying me for hacking on the VM so I have no obligation to
anybody. Ha!

Still, I want to return to the old times of regular updates and stuff.
It was fun and satisfying, and I want to get that back. Let's start with
something that should be easy - instruction encoder/decoder. It will
take some time to document what is actually needed, develop a format,
and implement it but that's OK. I think it should be quite refreshing to
actually achieve something, write some working code that will not
require serious thinking and theoretical deliberations. Simple things
for simple people, I guess...