New bytecode decoder
The new decoder has a much better API allowing for *huge* code savings
in client code (rougly 33% original line count is needed to decode the
instructions). Encoding structure is not changed, it remains exactly the
same, old, bloated encoding as it was... for now.

Now that we have the new simpler decoder and know what is needed to
actually decode the instructions we can work on the new *encoder*. And
then on a new assembler pipeline because, frankly, what we have now is
an embarassement.