Use size_t instead of typedef for schedulers and processes count
That type was a tad too big anyway, so there is no reason to keep it
typedefed. Just use size_t and be done with it. Unless the number of
schedulers is capped at some "weird" value (e.g. 42) then I see no
reason to try to be fancy with typedefs.