Fix: cancelling I/O works even if the fd never becomes ready
Before this fix cancellation was only detected in result of the work.
This meant that the interaction must have at least tried to do something
in order for the cancellation to be noticed. If the fd for the
interaction never became ready then the cancellation was never noticed
and the 'io_wait' waiting for the cancellation to take effect never
finished if it had timeout set to 'infinity' as it should.