Save exception throw points for easier debugging
Exceptions can now include information about "throw points". A throw
point is a location (bytecode address) at which the exception was
emitted. Failed program may be disassembled with --debug option to
discover the instruction that triggered the exception.

An example:

    stack trace: from entry point, most recent call last...
      main/0/0()

    uncaught object: Exception = no message received

    throw points:
        address: 0x0010 (byte 16) inside 0x55feedc34200 [a.out]
        address: 0x0033 (byte 51) inside 0x55feedc34200 [a.out]

    frame details:
      non-empty registers: 1/2:
        registers[1]: <Process> Process: 0x74deadbeef
      no arguments were passed to this frame

This means that the exception was first thrown at address 0x0010 inside
jump base 0x55feedc34200, which happens to be module "a.out". It was
then rethrown at address 0x0033 inside the same jump base.
If jump base addresses are the same it means that the exception did not
cross module boundaries.

If we disassemble the a.out module and check the bytes we can deduce the
chain of events that lead to the failure:

    $ viua-dis --debug a.out | less
    ...
        ; absolute address: 0x0010
        receive void 10ms
    ...
        ; absolute address: 0x0033
        join void %1 local infinity

The fact that the exception has more than one throw point suggests that
it was either caught but rethrown (maybe the handler was just adding
some additional information), or it was first thrown in a different
actor than the one that reported it.

With this information we can say that the actor did not receive a
message, did not cope well with this and died. Then, its parent actor
attempted to join it but did not expect an exception to arrive and also
died.