Fix: results are never recorded for crashed joinable processes
The race condition was triggered for processes that had a parent and
crashed as a result of an FFI call. The sequence of events was:

- FFI scheduler raises an exception and causes the stack to be unwound
- PROC scheduler sees that the process is suspended
- FFI scheduler wakes the process up
- PROC scheduler is past the suspend-check and goes to the top of the
  tick loop

The first condition to check in the ::tick() loop was if a process is
stopped (which was true), and if it is joinable (which was also true in
this case). If it was true the process was just pushed to the queue
without further processing... which meant that no result-recording was
performed and to the parent process it always looked like the failed
process is not joinable.

Removing the first condition in Process_scheduler::operator()() was
sufficient to fix the problem.