Fix: throw exceptions as exceptions not as generic value type
Throwing exceptions as generic values was causing C++ to use different
catch block and wrap the exception in another exception. This was not
what we wanted.

This commit fixes the bug. Exceptions are now thrown as exceptions, and
only values of other types are wrapped.

After this change the block catching generic values (viua::types::Value)
is no longer needed so the code was deleted. It is impossible for user
code (i.e. bytecode) to throw something that is not a proper exception.
FFI scheduler also wraps values in exceptions, and dispatches errors
directly so it is not using the same code path as bytecode parts of user
programs.