User std::set instead of std::vector to hold the set of pointers
What the value has is a *set* of pointers that are attached to it. The
order is not important - what is important is whether a pointer is
attached to the value (is present in the set), or is not attached to the
vaalue (is absent from the set). Nothing else is really needed, except
for inserting and erasing values from this set.