Expire values before crossing process boundary
This is to make pointers safer.

THE CURRENT SITUATION

Pointers are bound to Values, and Values know exactly what pointers are
bound to them. This machinery was put in place to allow pointers to know
when the value they point to is destroyed and they are no longer valid.
It works like this:

- a Pointer is created pointing to a Value
- the Value records the creation of the Pointer

...then, depending on the order of events (either the Value is destroyed
first, or the Pointer) one of two paths is taken. If the Value is
destroyed first:

- the Value runs its dtor
- the Value sends an expiration message to every Pointer that is
  attached to it
- each Pointer marks itself as expired
- the Value is destroyed

If the Pointer is destroyed first:

- the Pointer enters its dtor
- the Pointer detaches from the Value
- the Value removes the Pointer from its set of attached Pointers
- the Pointer is destroyed

THE PROBLEM

However, what happens if dtors of Value and Pointer are entered at the
same time? Bad things happen; memory corruption, for example.

Shouldn't a situation when dtors of Value and Pointer are entered at the
same time be impossible though - Pointers are invalidated when they
cross process boundaries so mutithreaded access seems extremely
unlikely...

Yes, it does seem extremely unlikely. But bugs lurk in extremely
unlikely places, where they are not looked for. The invalidation logic
was run on first dereference of the pointer. And what about the
situation when the Pointer is not dereferenced? The VM goes kaboom!

THE SOLUTION

OK, so let's just run invalidation upon receiving messages. Nope, too
late - execution of the Value's dtor may just started in the original
thread. We have to expire pointers before:

- they are sent to another process
- they are set as result of a process (as return values or exceptions
  may contain pointers)

Expiring values BEFORE they escape the original process will ensure
safety of access. Well, "will" is a bit strong here, so let's just say
"should".