Fix segfaults of Viuact lexer
I guess Viuact lexer was the first program big enough to really put some
stress on the VM. Unfortunately, it turned out that the stress was too
great and the VM couldn't handle it... nor could it handle the resuling
segmentation faults.

HOW DID THIS HAPPEN?

You know about something called "race condition"? Yep, it was two of
those little fuckers. The test suite was not good enough to catch it.

The sequence of events that triggered the first segfault was:

- a pointer is created
- later, the pointee (value that had the pointer taken) tries to
  invalidate the pointer
- the VM crashes

For some reason it resulting in either heap corruption, or later double
frees. I think that the issue was a bit more complex than what the list
above describes, but after I moved pointer tracking from pointer-value
into processes (so the pointer-pointee relationships have an external
arbiter) it went away.

The sequence of events that triggered the second segfault was:

- CPU starts executing a call instruction
- a foreign function F is called from process P
- process P gets suspended
- PROC scheduler requests a foreign function call from FFI scheduler
- PROC scheduler gets preempted by the OS
- FFI scheduler runs the foreign function F, but it emits an exception
- FFI scheduler sets the exception as thrown in process P
- FFI scheduler unwinds the stack of process P, which adjusts the
  instruction pointer to appropriate exception handler and moves the
  exception to caught slot
- FFI scheduler wakes process P up
- the OS runs PROC scheduler again
- PROC scheduler notices that an exception was thrown and unwinds the
  stack, effectively duplicating what the FFI scheduler already did

Due to the fact that the stack was unwound twice the exception that was
already in the caught slot was overwritten by a nullptr in the thrown
slot. How did it happen that the stack got unwound twice?

The code that executes a single instruction did not check for suspended
status immediately after it returned from an instruction implementation
(which it definitely should have done!). Now it does the correct thing
and no double-unwinding happens.

HOW WAS IT FIXED?

The first crash was fixed ogether with:

- 7d993fd5d0a830b2684fba280246dd91b625cf99 Expire values before crossing
  process boundary
- 4090fc9a241d785137d4d9f3f90d15d2f6d22f4a Let Process track pointers

The second one was fixed by this commit.