Drop old issues
After a long hiatus I am getting back to the VM for some time. It's
summer holidays so I can focus on my own projects instead of only doing
work-related research and stuff for students at the Uni.

I want to rework the VM. There are subsystems that are OK, but there are
others which are crumbling, creaking, and are generally in poor
condition. The issues that are dropped by this commit were created with
the assumption of continuous small patching, but since I will be
recreating the ISA almost from scratch they lost their relevance.

Bye old issues, you had a good run.