Use a different sigil for direct register access
Since the registers are not named in Viua VM's ISA, they are identified
by their indices. To distinguish them from ordinary integers in other
way than position after an instruction a sigil is used.

In the previous version the sigil character was % which is quite an
unfortunate choice as it is a symbol for an arithmetic operation. The
new sigil is $ which I believe is a better choice, as it does not
conflict with any mathematical notions.