Add a string ctor instruction: string
This will use an u64 from a register as an offset into strings table to
construct a boxed string "in" a register. Strings in the string table
are encoded as size-prefixed instead of zero-terminated. I believe this
is safer and more elegant; it also allows for embedding null characters
inside the strings.

This commit also introduces the root of the new type system, which is
needed now that the prototype has a notion of a boxed type instead of
only bit patterns to interpret as signed or unsigned integers.