Make room for tooling libraries
The source code should be arranged reasonably. What does this mean? It
means that source code for a specific subsystem of the VM (be it either
a scheduler implementation, or an assembler's frontend) should be placed
in a directory in which it would be reasonable to put it.

TL;DR Avoid making the codebase messy.

Now, how to do that is another matter entirely. My general idea is to
try to follow the following scheme:

- include/viua: put headers here, with directory structure matching
  namespace structure
- src/tools/exec: frontend implementations of various tools (assembler,
  linker, kernel, etc)
- src/tools/libs: backend implementations of tools, except the heart of
  the VM itself ie, the kernel
- src/arch: implementation of the ISA related functionality eg, encoding
  and decoding of the bytecode, stringification of bytecode; this should
  be mostly abstract and not related to concrete implementation of the
  VM or kernel
- src/vm: concrete implementation of the VM ie, schedulers (for CPU,
  I/O, and FFI), loader, stack and exception handling, message passing
  and distribution infrastructure, signal handling, error handling
- src/support: code that would be useful even in other projects; stuff
  like string manipulation, filesystem access, I/O interaction, etc

The future will tell how well this will work out.