Guard against the ip going out of bounds of .text section
The previous guard was just checking if the ip didn't reach the end of
the .text section. This was wholly inadequate because if the ip SKIPPED
the element at the end then the VM would start executing random garbage
from memory.

The new check is more comprehensive, and much better at its job than the
old one.