Rename framei opcode to frame
There is no "immediate" value used, so the "i" suffix was out of place.
The operand for call instruction is a register index, and depending on
what register SET is being indexed, the instruction treats the value
either as immediate (if the set was "argument"), or as a normal register
access (if it was anything else).