Improve handling of function table
Initially store function addresses as zero, just to be able to record
their offsets in the function table. Thanks to this level of indirection
we can always patch them later (the call instructions only need their
offsets in function table, not the real addresses).

Load the function table in the VM to see where each function is located.