Make the return value register explicit for return instructions
This saves just one instruction, but makes the source code more explicit
and immediately readable. Instead of:

    move $0.l, $42
    return

one can just write:

    return $42

Of course, this does not preclude using $0.l as the designated return
value register and keeping the old convention alive. As a matter of
fact, it's probably what I'm going to do.