Refactor bytecode emission
Move bytecode emission into a separate function. This prompted a slight
refactor of error repoting, since the ability of directly referencing
source text and file name from the bytecode-emitting function was lost.

The result of this need are the .start and .end member variables in the
Fn_def struct which define offsets in the source file at which the
function's definition starts and ends. They are used to translate the
offsets of errors thrown in the deep bowels of the assembler (ie, one
call frame deeper) back into names of functions inside which the error
was found. And displaying the name of the function which contains the
error is IMMENSELY useful for debugging.