Abstract the implementations of arithmetic operations with immmediates
It turns out, though, that it is not so simple. What should direct the
type of result? The instruction, or the value?

What about such an sequence of instructions:

    li $1, 42
    addiu $1, $1, 1

It is not unreasonable that the result should be an unsigned integer.
After all, no operands involved are negative.

Now, what about this:

    li $1, -42
    addiu $1, $1, 1

Should the result be signed or unsigned? The design of three-way
arithmetic operations dictates that the result is signed, because type
of the result is detemined by type of the left-hand side operand. This
makes sense.

However, what about this:

    li $1, -42
    addiu $1, $1, -1

Should it be outright illegal to use obviously signed immediates in
unsigned instructions? Yeah, probably.

Anyway, there are some interesting decisions to be made.