Mark unsigned opcodes with a bit flag
There are several arithmetic instructions which deal with unsigned
values. It would be useful to be able to detect them eg, for error
detection.

The assembler could detect that an operand is signed, and check if the
UNSIGNED bit in the opcode is high - if yes, then the instruction is
illegal.

Another useful feature of this change is that all "add" instructions are
encoded as 1, "sub" as 2, "mul" as 3, and "div" as 4. This makes opcodes
more consistent and elegant.