Crude implementations of eq, lt, gt instructions
Not the best or most elegant ones, but these will suffice until we have
something better. This is a prototype after all so experimenting is
explicitly permitted, encouraged even. Comparisons will stay hideous
until I devise a better way of implementing them. The trait system is a
little bit... how to put this... damaged? Let's hope I get a better idea
how to implement it soon.