Add g.li pseudoinstruction
The g.li pseudoinstruction is exactly the same as the li
pseudoinstruction, but is emitted in such a way that the last
instruction in emitted sequence has the greedy bit set.

It is useful for opcodes such as string, float, or double and allows
emitting long, greedy sequences for loading them. This is consistent
with how li itself is expanded - with a chain of greedy ops.