Reduce the number of opcodes visible to programmers
Remove the unsigned variants of immediate arithmetic instructions. In
order to perform unsigned arithmetic use unsigned integer as the
immediate operand. For example, to store unsigned 42 in a register, the
following code should be used:

    addi $1, void, 42u

Instead of the following, which was used BEFORE this patch:

    addiu $1, void, 42

This makes the public interface (ie, the instructions visible to
programmers) more consistent - the li instruction also uses the
signedness of the operand to determine whether to load signed or
unsigned integer.